Refactor search, filter and media stores
