+++ src/components/NavSection.vue
@@ -125,10 +125,10 @@
 </template>
 
 <script>
-import { SET_Q } from '~/constants/mutation-types'
+import { SET_Q } from '~/constants/action-types'
 import Dropdown from '~/components/Dropdown'
 import { SEARCH } from '~/constants/store-modules'
-import { mapMutations } from 'vuex'
+import { mapActions } from 'vuex'
 
 export default {
   name: 'NavSection',
@@ -145,7 +145,7 @@ export default {
     },
   },
   methods: {
-    ...mapMutations(SEARCH, { setSearchTerm: SET_Q }),
+    ...mapActions(SEARCH, { setSearchTerm: SET_Q }),
     onSubmit() {
       const q = this.form.searchTerm
       const theArgs = { q }
