+++ src/components/PhotoTags.vue
@@ -19,9 +19,9 @@
 </template>
 
 <script>
-import { SET_Q } from '~/constants/mutation-types'
+import { SET_Q } from '~/constants/action-types'
 import { SEARCH } from '~/constants/store-modules'
-import { mapMutations } from 'vuex'
+import { mapActions } from 'vuex'
 
 export default {
   name: 'PhotoTags',
@@ -32,7 +32,7 @@ export default {
     },
   },
   methods: {
-    ...mapMutations(SEARCH, { setSearchTerm: SET_Q }),
+    ...mapActions(SEARCH, { setSearchTerm: SET_Q }),
     isClarifaiTag(provider) {
       return provider === 'clarifai'
     },
