+++ src/constants/action-types.js
@@ -15,3 +15,6 @@ export const HANDLE_NO_MEDIA = 'HANDLE_NO_MEDIA'
 export const NOTIFICATION_ACTION = 'NOTIFICATION_ACTION'
 export const UPDATE_QUERY = 'UPDATE_QUERY'
 export const SET_FILTERS_FROM_URL = 'SET_FILTERS_FROM_URL'
+export const REPLACE_QUERY = 'REPLACE_QUERY'
+export const SET_QUERY = 'SET_QUERY'
+export const SET_Q = 'SET_Q'
