+++ src/constants/mutation-types.js
@@ -14,7 +14,6 @@ export const SET_IMAGE_PAGE = 'SET_IMAGE_PAGE'
 export const SET_MEDIA_PROVIDERS = 'SET_MEDIA_PROVIDERS'
 export const SET_PROVIDERS_FILTERS = 'SET_PROVIDERS_FILTERS'
 export const SET_MEDIA = 'SET_MEDIA'
-export const SET_QUERY = 'SET_QUERY'
 export const REPORT_BUG_START = 'REPORT_BUG_START'
 export const REPORT_BUG_END = 'REPORT_BUG_END'
 export const REPORT_BUG_FAILED = 'REPORT_BUG_FAILED'
@@ -34,5 +33,4 @@ export const PAUSE_ACTIVE_MEDIA_ITEM = 'PAUSE_ACTIVE_MEDIA_ITEM'
 export const EJECT_ACTIVE_MEDIA_ITEM = 'EJECT_ACTIVE_MEDIA_ITEM'
 export const RESET_MEDIA = 'RESET_MEDIA'
 export const REPLACE_FILTERS = 'REPLACE_FILTERS'
-export const REPLACE_QUERY = 'REPLACE_QUERY'
-export const SET_Q = 'SET_Q'
+export const MUTATE_QUERY = 'MUTATE_QUERY'
