+++ src/constants/store-modules.js
@@ -7,6 +7,6 @@ export const NAV = 'nav'
 export const NOTIFICATION = 'notification'
 export const REPORT_CONTENT = 'report-content'
 export const USAGE_DATA = 'usage-data'
-export const FILTER = 'filter'
 export const SEARCH = 'search'
+export const MEDIA = 'media'
 export const BUG_REPORT = 'bug-report'
