+++ src/store/provider.js
@@ -14,7 +14,7 @@ import {
   SET_MEDIA_PROVIDERS,
   SET_PROVIDERS_FILTERS,
 } from '~/constants/mutation-types'
-import { FILTER } from '~/constants/store-modules'
+import { SEARCH } from '~/constants/store-modules'
 
 const AudioProviderService = MediaProviderService(AUDIO)
 const ImageProviderService = MediaProviderService(IMAGE)
@@ -73,7 +73,7 @@ export const createActions = (services) => ({
           providers: sortedProviders,
         })
         commit(
-          `${FILTER}/${SET_PROVIDERS_FILTERS}`,
+          `${SEARCH}/${SET_PROVIDERS_FILTERS}`,
           {
             mediaType,
             providers: sortedProviders,
