+++ src/utils/get-legacy-source-url.js
@@ -32,7 +32,7 @@ export const legacySourceMap = {
       let query = `${search.q} AND RIGHTS:*creative*` // search cc licensed works
 
       if (search.filters && search.filters.commercial) {
-        if (search.filters.comercial) query = `${query} AND NOT RIGHTS:*nc*`
+        if (search.filters.commercial) query = `${query} AND NOT RIGHTS:*nc*`
         if (search.filters.modify) query = `${query} AND NOT RIGHTS:*nd*`
       }
