+++ src/utils/search-query-transform.js
@@ -1,6 +1,6 @@
 import clonedeep from 'lodash.clonedeep'
 import findIndex from 'lodash.findindex'
-import { filterData, mediaSpecificFilters } from '~/store/filter'
+import { filterData, mediaSpecificFilters } from '~/store/search'
 import getParameterByName from './get-parameter-by-name'
 import { ALL_MEDIA, IMAGE } from '~/constants/media'
