+++ test/unit/specs/utils/search-query-transform.spec.js
@@ -4,7 +4,7 @@ import {
   queryToFilterData,
   queryStringToQueryData,
 } from '~/utils/search-query-transform'
-import { filterData } from '~/store/filter'
+import { filterData } from '~/store/search'
 import { IMAGE } from '~/constants/media'
 
 describe('searchQueryTransform', () => {
